django-web-app/merchex/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==5.0.6
django-appconf==1.0.6
django-cloudinary-storage==0.3.0
django-imagekit==5.0.0
idna==3.7
numpy==2.0.0
opencv-python==4.10.0.84
pilkit==3.0
Pillow==10.0.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
